Create empty state component
